@@ -1,11 +1,11 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: prefer_final_fields, public_member_api_docs, prefer_single_quotes, omit_local_variable_types, unnecessary_this
+// ignore_for_file: public_member_api_docs, prefer_single_quotes, avoid_escaping_inner_quotes, prefer_const_constructors, sort_constructors_first, always_specify_types
 
 import 'dart:async';
 
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
-/// A  class generated by flappy_translator package containing localized strings
+/// A class generated by flappy_translator package containing localized strings
 class AppLocalizations {
   static String get appTitle => _getText('appTitle');
 
@@ -40,9 +40,9 @@ class AppLocalizations {
   static String get gameAlreadyOpenPopupDescription =>
       _getText('gameAlreadyOpenPopupDescription');
 
-  static Map<String, String> _localizedValues;
+  static late Map<String, String> _localizedValues;
 
-  static Map<String, String> _enValues = {
+  static const _enValues = {
     'appTitle': 'Kifu Viewer',
     'generalOk': 'Ok',
     'generalOpen': 'Open',
@@ -61,20 +61,19 @@ class AppLocalizations {
         'Game already open. Please choose another game.',
   };
 
-  static Map<String, Map<String, String>> _allValues = {
+  static const _allValues = {
     'en': _enValues,
   };
 
   AppLocalizations(Locale locale) {
     _locale = locale;
-    _localizedValues = null;
+    _localizedValues = {};
   }
 
-  static Locale _locale;
+  static late Locale _locale;
 
-  static String _getText(String key) {
-    return _localizedValues[key] ?? '** $key not found';
-  }
+  static String _getText(String key) =>
+      _localizedValues[key] ?? '** $key not found';
 
   static Locale get currentLocale => _locale;
 
@@ -82,7 +81,7 @@ class AppLocalizations {
 
   static Future<AppLocalizations> load(Locale locale) async {
     final translations = AppLocalizations(locale);
-    _localizedValues = _allValues[locale.toString()];
+    _localizedValues = _allValues[locale.toString()]!;
     return translations;
   }
 }
